update install directions
Tried to make the directions/commands for installing and startup more explicit. I don't know if the old `code` snippets were supposed to be suggestive or platform specific, but they don't make sense if you're on mac/unix.

All these commands work for me in the terminal on a mac. The only thing I skipped is setting up a virtual environment with `virtualenv`, but that could be another section and I didn't want to assume much for the basic setup.

For the dependencies, I listed everything in `INSTALL.md` instead of just NumPy.

Made a few grammar adjustments as well.